refactor(sql-validator): sql_validator.py를 4개 모듈로 분할하고 SYSTEM_SCHEMAS 단일 소스화
- sql_identifier_utils.py: 별칭/CTE/파생 테이블 파싱 유틸리티 이동 (CC-002)
- sql_metadata.py: SchemaMetadata/SchemaMetadataProvider 이동, FUZZY_MATCH_CUTOFF 상수 도입 (CC-002, CC-016)
- sql_autocompleter.py: SQLAutoCompleter 이동, get_completions을 4개 사설 헬퍼로 분해 (CC-002, CC-010)
- sql_validator.py: IssueSeverity/ValidationIssue/SQLValidator만 남기고 하위호환 재수출 블록 추가,
  _validate_version_compatibility의 중복 스캔 루프를 _flag_unsupported_items로 통합 (CC-009)
- constants.py: SYSTEM_SCHEMAS에 'ndbinfo' 추가, SQLValidator.SYSTEM_SCHEMAS를 constants 파생 상수로 교체 (CC-008)
- 순환참조 방지를 위해 sql_metadata를 별도 모듈로 분리해 sql_identifier_utils/sql_metadata → sql_validator/sql_autocompleter 단방향 의존 구조 확정
- 기존 소비자 6곳(src/core/__init__.py 등)은 재수출 유지로 무수정
- 재수출 identity 가드 + SYSTEM_SCHEMAS 파생 가드 테스트 추가
